Replace use of off_t by size_t in cparser

+++ astropy/io/ascii/src/tokenizer.c
@@ -410,7 +410,7 @@ int tokenize(tokenizer_t *self, int end, int header, int num_cols)
                 {
                     // In this case we don't want to left-strip the field,
                     // so we backtrack
-                    off_t tmp = self->source_pos;
+                    size_t tmp = self->source_pos;
                     --self->source_pos;
 
                     while (self->source_pos >= 0 &&
